+++ apps/backend/src/app.module.ts
@@ -2,6 +2,8 @@ import { Module } from '@nestjs/common';
 import { ConfigModule } from '@nestjs/config';
 import { AuthModule } from './modules/auth/auth.module';
 import { UsersModule } from './modules/users/users.module';
+import { BriefingsModule } from './modules/briefings/briefings.module';
+import { ProposalsModule } from './modules/proposals/proposals.module';
 import { PrismaModule } from './prisma/prisma.module';
 
 @Module({
@@ -13,6 +15,8 @@ import { PrismaModule } from './prisma/prisma.module';
     PrismaModule,
     AuthModule,
     UsersModule,
+    BriefingsModule,
+    ProposalsModule,
   ],
 })
 export class AppModule {}